requirements.txt
absl-py==2.0.0
aiohttp==3.8.6
aiosignal==1.3.1
altair==5.1.2
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.3
boxmot==10.0.43
cachetools==5.3.2
certifi==2023.7.22
cfgv==3.4.0
charset-normalizer==3.3.1
click==8.1.7
coloredlogs==15.0.1
contourpy==1.1.1
cycler==0.12.1
Cython==3.0.4
distlib==0.3.7
exceptiongroup==1.1.3
fastapi==0.104.0
filelock==3.13.0
filterpy==1.4.5
flatbuffers==23.5.26
fonttools==4.43.1
frozenlist==1.4.0
fsspec==2023.10.0
ftfy==6.1.1
gdown==4.7.1
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.23.3
google-auth-oauthlib==1.1.0
grpcio==1.59.0
h11==0.14.0
humanfriendly==10.0
hyperlpr3==0.1.3
identify==2.5.31
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
lap==0.4.0
lapx==0.5.5
loguru==0.7.2
Markdown==3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
networkx==3.2
nodeenv==1.8.0
numpy==1.23.1
oauthlib==3.2.2
omegaconf==2.3.0
onnx==1.15.0
onnxruntime==1.16.1
openalpr==1.1.0
opencv-python==4.8.1.78
packaging==23.2
pandas==2.1.2
Pillow==10.1.0
platformdirs==3.11.0
pre-commit==3.5.0
protobuf==4.23.4
psutil==5.9.6
psycopg2==2.9.9
psycopg2-binary==2.9.9
py-cpuinfo==9.0.0
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.4.2
pydantic_core==2.10.1
pydeck==0.8.1b0
Pygments==2.16.1
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.6.0
rpds-py==0.10.6
rsa==4.9
scikit-learn==1.3.2
scipy==1.11.3
seaborn==0.13.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
starlette==0.27.0
streamlit==1.28.0
sympy==1.12
tenacity==8.2.3
tensorboard==2.15.0
tensorboard-data-server==0.7.2
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
toml==0.10.2
toolz==0.12.0
torch==2.1.0
torchvision==0.16.0
tornado==6.3.3
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
ultralytics==8.0.202
urllib3==2.0.7
uvicorn==0.23.2
validators==0.22.0
virtualenv==20.24.6
wcwidth==0.2.8
Werkzeug==3.0.1
yacs==0.1.8
yarl==1.9.2
zipp==3.17.0